+++ core/src/scripting.rs
@@ -1,4 +1,7 @@
-use crate::{modules::{make_env_module, make_fs_module, make_shell_module, make_proc_module, make_modules}, util::resolve_path};
+use crate::{
+    modules::{make_env_module, make_fs_module, make_modules, make_proc_module, make_shell_module},
+    util::resolve_path,
+};
 use log::info;
 use mlua::{Error as LuaError, Lua};
 use std::{